update test suite to test flattened cols from dataframe
