--- spark/src/test/scala/io/spark/ddf/NestedDDFSuite.scala
@@ -1,31 +1,0 @@
-package io.spark.ddf
-
-
-import org.apache.spark.sql.{SQLContext, DataFrame}
-
-/**
-  */
-class NestedDDFSuite extends ATestSuite {
-
-  /*
-  test("test JSON with primitiveFieldAndType") {
-
-    val df:DataFrame = jsonRDD(primitiveFieldAndType)
-
-    df.printSchema()
-    println("Number of records: " + df.count())
-
-  }
-  */
-  test("Loading a JSON dataset from a text file") {
-    val path = "resources/test/sleep_data_sample.json"
-    val sqlCtx = manager.getHiveContext
-    val df:DataFrame = sqlCtx.jsonFile(path)
-
-    println("dataframe created from json file at " + path)
-    df.printSchema()
-    println("Number of records: " + df.count())
-
-  }
-
-}
